// src/app/api/models/email-settings-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface EmailSettingsDto {
  advancedSettings?: {
};
  fromEmail: string;
  fromName: string;
  isActive: boolean;
  provider: 'gmail' | 'outlook' | 'yahoo' | 'custom';
  replyToEmail?: string;
  smtpHost: string;
  smtpPassword: string;
  smtpPort: number;
  smtpSecure: boolean;
  smtpUser: string;
}