// src/app/api/models/email-settings-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface EmailSettingsResponseDto {
  advancedSettings?: {
};
  createdAt: string;
  fromEmail: string;
  fromName: string;
  id: string;
  isActive: boolean;
  lastTestedAt?: string;
  provider: 'gmail' | 'outlook' | 'yahoo' | 'custom';
  replyToEmail?: string;
  smtpHost: string;
  smtpPassword: string;
  smtpPort: number;
  smtpSecure: boolean;
  smtpUser: string;
  testStatus?: 'success' | 'failed';
  updatedAt: string;
}